+++ changelog.md
@@ -2,6 +2,8 @@
 
 **Version 0.2.0**
 
+* `save_as` was removed from `train`. Use ModelCheckpoint instead and provide it as a callback. 
+
 * `show_images` now shows titles. These functions will be removed later and imported from `pyautocv`
 instead.